+++ src/counter_analysis_toolkit/prepareArray.c
@@ -7,15 +7,15 @@
 #include "prepareArray.h"
 
 volatile uintptr_t opt_killer_zero;
-static void _prepareArray_sections_random(uintptr_t *array, int len, int stride, long secSize);
-static void _prepareArray_sequential(uintptr_t *array, int len, int stride);
+static void _prepareArray_sections_random(uintptr_t *array, long long len, long long stride, long long secSize);
+static void _prepareArray_sequential(uintptr_t *array, long long len, long long stride);
 
 
 /*
  * "stride" is in "uintptr_t" elements, NOT in bytes
  * Note: It is wise to provide an "array" that is aligned to the cache line size.
  */
-int prepareArray(uintptr_t *array, int len, int stride, long secSize, int pattern){
+int prepareArray(uintptr_t *array, long long len, long long stride, long long secSize, int pattern){
     assert( array != NULL );
     opt_killer_zero = (uintptr_t)( (len+37)/(len+36) - 1 );
 
@@ -38,21 +38,21 @@ int prepareArray(uintptr_t *array, int len, int stride, long secSize, int patter
  * "stride" is in "uintptr_t" elements, NOT in bytes
  * Note: It is wise to provide an "array" that is aligned to the cache line size.
  */
-static void _prepareArray_sections_random(uintptr_t *array, int len, int stride, long secSize){
+static void _prepareArray_sections_random(uintptr_t *array, long long len, long long stride, long long secSize){
 
     assert( array != NULL );
 
-    int elemCnt, maxElemCnt, sec, i;
-    int currElemCnt, uniqIndex, taken;
+    long long elemCnt, maxElemCnt, sec, i;
+    long long currElemCnt, uniqIndex, taken;
     uintptr_t **p, *next;
-    int currSecSize = secSize;
-    int secCnt = 1+len/secSize;
-    int *availableNumbers;
+    long long currSecSize = secSize;
+    long long secCnt = 1+len/secSize;
+    long long *availableNumbers;
 
     p = (uintptr_t **)&array[0];
 
     maxElemCnt = currSecSize/stride;
-    availableNumbers = (int *)calloc(maxElemCnt, sizeof(int));
+    availableNumbers = (long long *)calloc(maxElemCnt, sizeof(long long));
 
     // For every section of the array
     for(sec=0; sec<secCnt; ++sec){
@@ -68,10 +68,10 @@ static void _prepareArray_sections_random(uintptr_t *array, int len, int stride,
         taken = 0;
         if( 0==sec ) // For the first section we have already picked "0", so we must pick one less element.
             taken = 1;
-        int remainingElemCnt = currElemCnt;
+        long long remainingElemCnt = currElemCnt;
 
         for(elemCnt=0; elemCnt<currElemCnt-taken; ++elemCnt){
-            int index = taken + random() % (remainingElemCnt-taken); // skip the first "taken" elements
+            long long index = taken + random() % (remainingElemCnt-taken); // skip the first "taken" elements
             // For the first section we skip zero as a choice (we already selected that before the loop.)
             uniqIndex = sec*secSize + stride*availableNumbers[index];
             // replace the chosen number with the last element.
@@ -98,8 +98,8 @@ static void _prepareArray_sections_random(uintptr_t *array, int len, int stride,
  * Note: It is wise to provide an "array" that is aligned to the cache line size.
  */
 
-static void _prepareArray_sequential(uintptr_t *array, int len, int stride){
-    int curr;
+static void _prepareArray_sequential(uintptr_t *array, long long len, long long stride){
+    long long curr;
     uintptr_t **p, *next;
 
     p = (uintptr_t **)&array[0];
